agent/requirements.txt
livekit-agents>=1.2.0
livekit-plugins-openai>=1.0.0
livekit-plugins-deepgram>=1.0.0
livekit-plugins-silero>=1.0.0
python-dotenv>=1.0.0
